refactor(draw): move the draw functions into there own file
